refactor: comment-api hook add
