+++ src/components/features/contents/comment/comment.tsx
@@ -1,11 +1,10 @@
 import SubTitle from '@components/common/text/SubTitle';
-import { useCallback, useEffect, useState } from 'react';
 import { useParams } from 'react-router-dom';
 
-import { ICommentAdd, useCommentAdd, useCommentById } from '@/api/comment';
 import Divider from '@/components/common/divider/divider';
 import Input from '@/components/common/Input/Input';
 import CommentContainer from '@/components/features/contents/comment/commentContainer';
+import useComment from '@/hooks/useComment';
 
 import * as S from '../style';
 
@@ -15,31 +14,8 @@ export interface IButtonData {
 }
 const ContentContainer = () => {
     const { id } = useParams();
-    const { data: commentsData } = useCommentById(Number(id));
-    const [inputValue, setinputValue] = useState('');
-    const handleValueChange = useCallback((value: string) => {
-        setinputValue(value);
-    }, []);
 
-    const { mutate: addComment, isError, isSuccess } = useCommentAdd();
-
-    /**
-     * react-query mutation으로 query-key update
-     */
-    const handleCommentSubmit = async () => {
-        const obj: ICommentAdd = {
-            boardId: Number(id),
-            content: inputValue,
-        };
-
-        addComment(obj);
-    };
-
-    useEffect(() => {
-        if (!isError && isSuccess) {
-            setinputValue('');
-        }
-    }, [isError, isSuccess]);
+    const { commentsData, inputValue, handleValueChange, handleCommentSubmit } = useComment(id);
 
     return (
         <S.Wrapper>
